+++ podcasts/Bookmarks/Player/BookmarksPlayerTabController.swift
@@ -73,6 +73,13 @@ class BookmarksPlayerTabController: PlayerItemViewController {
     }
 
     private func handleBookmarkCreated(bookmark: Bookmark) {
+        // Prevent the add bookmark window from opening if the app isn't active
+        // We also prevent it from opening while connected to CarPlay to not distract anyone, and in my testing the app state is always
+        // true while connected to CarPlay, even if it's in the background
+        guard UIApplication.shared.applicationState == .active, !SceneHelper.isConnectedToCarPlay else {
+            return
+        }
+
         showBookmarkEdit(isNew: true, bookmark: bookmark)
     }